split worker and coordinator classes and fix idle status
